% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_barcode.R
\name{make_barcode_plot}
\alias{make_barcode_plot}
\title{Make DNA Barcode Plot}
\usage{
make_barcode_plot(sequences, species_sub, trim = FALSE)
}
\arguments{
\item{sequences}{DNAStringSet cnotaining sequences to be aligned and plotted.
Should be loaded from a FASTA file with \code{sequences <- Biostrings::readDNAStringSet('seqs.fasta')}}

\item{species_sub}{character vector, the names of the sequences to align and plot.
Must match `rownames(sequences)`}

\item{trim}{logical, should leading or trailing gaps be trimmed?
All sequences will be trimmed equally}
}
\value{
A ggplot containing a DNA barcode plot
}
\description{
Make DNA Barcode Plot
}
