% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_barcode.R
\name{make_dna_vectors}
\alias{make_dna_vectors}
\title{Make table of DNA sequence for graphing}
\usage{
make_dna_vectors(seqs, trim = FALSE)
}
\arguments{
\item{seqs}{DNAStringSet containing sequences to align}

\item{trim}{logical, should leading or trailing gaps be trimmed?
All sequences will be trimmed equally}
}
\value{
A data.frame in long-format, with 1 row per sequence and base
}
\description{
Make table of DNA sequence for graphing
}
